# webapp/frontend/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "ComparisonDetail": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "git_sha": {
            "title": "Git Sha",
            "type": "string"
          },
          "launched_by_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Launched By Username"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "per_strategy_stats": {
            "items": {
              "$ref": "#/components/schemas/PerStrategyStatsRow"
            },
            "title": "Per Strategy Stats",
            "type": "array"
          },
          "plots": {
            "items": {
              "type": "string"
            },
            "title": "Plots",
            "type": "array"
          },
          "store": {
            "title": "Store",
            "type": "string"
          }
        },
        "required": [
          "name",
          "store",
          "created_at",
          "git_sha",
          "per_strategy_stats",
          "plots"
        ],
        "title": "ComparisonDetail",
        "type": "object"
      },
      "ComparisonSummary": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "launched_by_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Launched By Username"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "store": {
            "title": "Store",
            "type": "string"
          },
          "strategies": {
            "items": {
              "type": "string"
            },
            "title": "Strategies",
            "type": "array"
          }
        },
        "required": [
          "name",
          "store",
          "created_at",
          "strategies"
        ],
        "title": "ComparisonSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PerStrategyStatsRow": {
        "properties": {
          "calmar_ci95_high": {
            "title": "Calmar Ci95 High",
            "type": "number"
          },
          "calmar_ci95_low": {
            "title": "Calmar Ci95 Low",
            "type": "number"
          },
          "calmar_mean": {
            "title": "Calmar Mean",
            "type": "number"
          },
          "calmar_std": {
            "title": "Calmar Std",
            "type": "number"
          },
          "experiment_id": {
            "title": "Experiment Id",
            "type": "string"
          },
          "max_drawdown_mean": {
            "title": "Max Drawdown Mean",
            "type": "number"
          },
          "max_drawdown_worst": {
            "title": "Max Drawdown Worst",
            "type": "number"
          },
          "n_folds": {
            "title": "N Folds",
            "type": "integer"
          },
          "sharpe_ci95_high": {
            "title": "Sharpe Ci95 High",
            "type": "number"
          },
          "sharpe_ci95_low": {
            "title": "Sharpe Ci95 Low",
            "type": "number"
          },
          "sharpe_mean": {
            "title": "Sharpe Mean",
            "type": "number"
          },
          "sharpe_std": {
            "title": "Sharpe Std",
            "type": "number"
          },
          "sortino_ci95_high": {
            "title": "Sortino Ci95 High",
            "type": "number"
          },
          "sortino_ci95_low": {
            "title": "Sortino Ci95 Low",
            "type": "number"
          },
          "sortino_mean": {
            "title": "Sortino Mean",
            "type": "number"
          },
          "sortino_std": {
            "title": "Sortino Std",
            "type": "number"
          },
          "strategy": {
            "title": "Strategy",
            "type": "string"
          },
          "total_return_mean": {
            "title": "Total Return Mean",
            "type": "number"
          },
          "total_return_std": {
            "title": "Total Return Std",
            "type": "number"
          },
          "trade_count_total": {
            "title": "Trade Count Total",
            "type": "integer"
          },
          "win_rate_mean": {
            "title": "Win Rate Mean",
            "type": "number"
          }
        },
        "required": [
          "strategy",
          "experiment_id",
          "n_folds",
          "sharpe_mean",
          "sharpe_std",
          "sharpe_ci95_low",
          "sharpe_ci95_high",
          "sortino_mean",
          "sortino_std",
          "sortino_ci95_low",
          "sortino_ci95_high",
          "calmar_mean",
          "calmar_std",
          "calmar_ci95_low",
          "calmar_ci95_high",
          "total_return_mean",
          "total_return_std",
          "max_drawdown_mean",
          "max_drawdown_worst",
          "win_rate_mean",
          "trade_count_total"
        ],
        "title": "PerStrategyStatsRow",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Quant Webapp",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/comparisons": {
      "get": {
        "operationId": "get_comparisons_api_comparisons_get",
        "parameters": [
          {
            "in": "query",
            "name": "all",
            "required": false,
            "schema": {
              "default": false,
              "title": "All",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ComparisonSummary"
                  },
                  "title": "Response Get Comparisons Api Comparisons Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Comparisons",
        "tags": [
          "comparisons"
        ]
      }
    },
    "/api/comparisons/{name}": {
      "get": {
        "operationId": "get_comparison_detail_api_comparisons__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComparisonDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Comparison Detail",
        "tags": [
          "comparisons"
        ]
      }
    },
    "/api/comparisons/{name}/plots/{plot_name}": {
      "get": {
        "operationId": "get_comparison_plot_api_comparisons__name__plots__plot_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "plot_name",
            "required": true,
            "schema": {
              "title": "Plot Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Comparison Plot",
        "tags": [
          "comparisons"
        ]
      }
    },
...
  }
}